refactored file structor and changed mysql user to roots
