+++ client/__test__/app.test.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import VideoPlayer from '../videoPlayer';
+import App from '../components/app';
+
 
 afterEach(cleanup);
 
 it('matches snapshot', () => {
-  const { asFragment } = render(<VideoPlayer link='https://www.youtube.com/embed/2LNOT2TYcmM' id='1' />);
+  const { asFragment } = render(<App />);
   expect(asFragment()).toMatchSnapshot();
+
 });
