+++ client/__test__/videoPlayer.test.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import App from '../app';
-
+import VideoPlayer from '../components/videoPlayer';
 
 afterEach(cleanup);
 
 it('matches snapshot', () => {
-  const { asFragment } = render(<App />);
+  const { asFragment } = render(<VideoPlayer link='https://www.youtube.com/embed/2LNOT2TYcmM' id='1' />);
   expect(asFragment()).toMatchSnapshot();
-
 });
