+++ client/components/details.js
