+++ client/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app/app';
+import App from './components/app';
 
 ReactDOM.render(<App />, document.getElementById('Details'));
